switch to vineflower decompilation

+++ spire_src/com/badlogic/gdx/tools/flame/Slider.java
@@ -1,6 +1,3 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package com.badlogic.gdx.tools.flame;
 
 import java.awt.BorderLayout;
@@ -10,34 +7,32 @@ import javax.swing.JSpinner;
 import javax.swing.SpinnerNumberModel;
 import javax.swing.event.ChangeListener;
 
-public class Slider
-extends JPanel {
-    public JSpinner spinner;
+public class Slider extends JPanel {
+   public JSpinner spinner;
 
-    public Slider(float initialValue, float min, float max, float stepSize) {
-        this.spinner = new JSpinner(new SpinnerNumberModel(initialValue, min, max, stepSize));
-        this.setLayout(new BorderLayout());
-        this.add(this.spinner);
-    }
+   public Slider(float initialValue, float min, float max, float stepSize) {
+      this.spinner = new JSpinner(new SpinnerNumberModel((double)initialValue, (double)min, (double)max, (double)stepSize));
+      this.setLayout(new BorderLayout());
+      this.add(this.spinner);
+   }
 
-    public void setValue(float value) {
-        this.spinner.setValue(value);
-    }
+   public void setValue(float value) {
+      this.spinner.setValue((double)value);
+   }
 
-    public float getValue() {
-        return ((Double)this.spinner.getValue()).floatValue();
-    }
+   public float getValue() {
+      return ((Double)this.spinner.getValue()).floatValue();
+   }
 
-    public void addChangeListener(ChangeListener listener) {
-        this.spinner.addChangeListener(listener);
-    }
+   public void addChangeListener(ChangeListener listener) {
+      this.spinner.addChangeListener(listener);
+   }
 
-    @Override
-    public Dimension getPreferredSize() {
-        Dimension size = super.getPreferredSize();
-        size.width = 75;
-        size.height = 26;
-        return size;
-    }
+   @Override
+   public Dimension getPreferredSize() {
+      Dimension size = super.getPreferredSize();
+      size.width = 75;
+      size.height = 26;
+      return size;
+   }
 }
-
