+++ spire_src/com/megacrit/cardcrawl/actions/unique/AddCardToDeckAction.java
@@ -1,6 +1,3 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package com.megacrit.cardcrawl.actions.unique;
 
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
@@ -9,19 +6,17 @@ import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.vfx.cardManip.ShowCardAndObtainEffect;
 
-public class AddCardToDeckAction
-extends AbstractGameAction {
-    AbstractCard cardToObtain;
+public class AddCardToDeckAction extends AbstractGameAction {
+   AbstractCard cardToObtain;
 
-    public AddCardToDeckAction(AbstractCard card) {
-        this.cardToObtain = card;
-        this.duration = 0.5f;
-    }
+   public AddCardToDeckAction(AbstractCard card) {
+      this.cardToObtain = card;
+      this.duration = 0.5F;
+   }
 
-    @Override
-    public void update() {
-        AbstractDungeon.effectList.add(new ShowCardAndObtainEffect(this.cardToObtain, (float)Settings.WIDTH / 2.0f, (float)Settings.HEIGHT / 2.0f));
-        this.isDone = true;
-    }
+   @Override
+   public void update() {
+      AbstractDungeon.effectList.add(new ShowCardAndObtainEffect(this.cardToObtain, Settings.WIDTH / 2.0F, Settings.HEIGHT / 2.0F));
+      this.isDone = true;
+   }
 }
-
