+++ spire_src/com/megacrit/cardcrawl/actions/watcher/SanctityAction.java
@@ -1,6 +1,3 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package com.megacrit.cardcrawl.actions.watcher;
 
 import com.badlogic.gdx.graphics.Color;
@@ -13,23 +10,24 @@ import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.vfx.BorderFlashEffect;
 import com.megacrit.cardcrawl.vfx.combat.SanctityEffect;
 
-public class SanctityAction
-extends AbstractGameAction {
-    private int amtToDraw;
+public class SanctityAction extends AbstractGameAction {
+   private int amtToDraw;
 
-    public SanctityAction(int amount) {
-        this.amtToDraw = amount;
-    }
+   public SanctityAction(int amount) {
+      this.amtToDraw = amount;
+   }
 
-    @Override
-    public void update() {
-        if (AbstractDungeon.actionManager.cardsPlayedThisCombat.size() >= 2 && AbstractDungeon.actionManager.cardsPlayedThisCombat.get((int)(AbstractDungeon.actionManager.cardsPlayedThisCombat.size() - 2)).type == AbstractCard.CardType.SKILL) {
-            this.addToTop(new DrawCardAction(this.amtToDraw));
-            this.addToTop(new VFXAction(new SanctityEffect(AbstractDungeon.player.hb.cX, AbstractDungeon.player.hb.cY)));
-            this.addToTop(new SFXAction("HEAL_1"));
-            this.addToTop(new VFXAction(new BorderFlashEffect(Color.GOLD, true), 0.1f));
-        }
-        this.isDone = true;
-    }
+   @Override
+   public void update() {
+      if (AbstractDungeon.actionManager.cardsPlayedThisCombat.size() >= 2
+         && AbstractDungeon.actionManager.cardsPlayedThisCombat.get(AbstractDungeon.actionManager.cardsPlayedThisCombat.size() - 2).type
+            == AbstractCard.CardType.SKILL) {
+         this.addToTop(new DrawCardAction(this.amtToDraw));
+         this.addToTop(new VFXAction(new SanctityEffect(AbstractDungeon.player.hb.cX, AbstractDungeon.player.hb.cY)));
+         this.addToTop(new SFXAction("HEAL_1"));
+         this.addToTop(new VFXAction(new BorderFlashEffect(Color.GOLD, true), 0.1F));
+      }
+
+      this.isDone = true;
+   }
 }
-
